# Optimize nullable reference type handling