--- a/azure-pipelines.yml
+++ b/azure-pipelines.yml
@@ -20,7 +20,7 @@ steps:
   displayName: 'Install .NET Core SDK'
   inputs:
     packageType: 'sdk'
-    version: '5.0.102'
+    version: '6.0.100'
 
 # Patch preview project versions (only when on master branch)
 - task: CmdLine@2