--- a/src/Namotion.Reflection.Benchmark/Namotion.Reflection.Benchmark.csproj
+++ b/src/Namotion.Reflection.Benchmark/Namotion.Reflection.Benchmark.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netcoreapp5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <LangVersion>9</LangVersion>
     <Nullable>enable</Nullable>
   </PropertyGroup>