`GetFlagsFromCustomAttributeProviders` showed up in NSwag generation as really hot method (and taking a lot of time). This accumulates as the information is loaded per type and all its members.

The old logic basically caused always search to member's type and assembly, using `GetCustomAttributes`, using dynamic object. 

Now:

* only using dynamic on old runtimes, other ones now use `ICustomAttributeProvider`
* using separate struct to store resolution result, caching result per type / assembly
   * method/property resolution is still done, but fallback to type/assembly won't be so costly
* better cache keys that do not require string concatenation on lookup (C# 10 feature)
* updated build to happen on NET 6 SDK

Tested the chain and no changes in client output for our nullable reference type -enabled project.

## Benchmark

Need to clear the cache to show the first hit problem.

```csharp
[MemoryDiagnoser]
public class ContextualPropertyBenchmark
{
    [Benchmark]
    public void GetContextualProperties()
    {
        CachedType.ClearCache();
        typeof(JsonSchema).GetContextualProperties();
    }
}
```

## Results

``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.100
  [Host]     : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  DefaultJob : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
```

### Before

|                  Method |     Mean |     Error |    StdDev |   Gen 0 | Allocated |
|------------------------ |---------:|----------:|----------:|--------:|----------:|
| GetContextualProperties | 1.755 ms | 0.0283 ms | 0.0265 ms | 95.7031 |    394 KB |


### After

|                  Method |     Mean |   Error |  StdDev |   Gen 0 |  Gen 1 | Allocated |
|------------------------ |---------:|--------:|--------:|--------:|-------:|----------:|
| GetContextualProperties | 306.1 μs | 5.65 μs | 5.01 μs | 20.5078 | 0.9766 |     86 KB |