@lahma 2021-12-05
With this change our client generation goes from `Duration: 00:00:07.7439235` to `Duration: 00:00:06.8724467`